requirements.txt

# pyrogram
# TgCrypto
# python-dotenv


pyrogram==2.0.106
tgcrypto==1.2.5
py-tgcalls==0.9.7
python-dotenv==1.0.1